+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - When removing rows while creating substance parameters, the entires where only removed from the gui, but not from the file.
 - Now uses Farthest Point Search for initial recommendation in discrete searchspaces
+- An error window is opened if a parameter name is already in use instead of passing silently
 ## [0.2.0] - 2025-10-04
 ### Added
 - Entires in the recommendation table now get checked against the allowed values